# Backend_Jobaile/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jobaile API",
        "description": "Dokumentasi API untuk aplikasi Jobaile (versi pertama)",
        "contact": {
            "name": "Jobaile Dev Team",
            "email": "support@jobaile.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/accbyworker/{id}": {
            "post": {
                "tags": [
                    "Matchmaking"
                ],
                "summary": "Worker menerima job offer",
                "operationId": "2430edfab180d7770301c11555413937",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID job offer yang ingin diterima",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job offer diterima, menunggu konfirmasi recruiter"
                    },
                    "401": {
                        "description": "Belum login / email belum diverifikasi"
                    },
                    "403": {
                        "description": "Hanya worker yang dapat menerima job"
                    },
                    "404": {
                        "description": "Job offer tidak valid atau sudah ditutup"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matchaccbyrecruiterorrequestbyrecruiter/{id_job}": {
            "post": {
                "tags": [
                    "Matchmaking"
                ],
                "summary": "Recruiter mengirimkan tawaran ke worker",
                "operationId": "dc84691027675985c6e9f9b5c2a89ae7",
                "parameters": [
                    {
                        "name": "id_job",
                        "in": "path",
                        "description": "ID dari job offer yang ingin ditawarkan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_worker"
                                ],
                                "properties": {
                                    "id_worker": {
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tawaran berhasil dikirim ke worker"
                    },
                    "401": {
                        "description": "Unauthorized atau email belum diverifikasi"
                    },
                    "403": {
                        "description": "Bukan recruiter"
                    },
                    "404": {
                        "description": "Job offer tidak ditemukan atau bukan milik recruiter"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/match/tolak/{id_job}": {
            "post": {
                "tags": [
                    "Matchmaking"
                ],
                "summary": "Tolak tawaran/match oleh worker atau recruiter",
                "operationId": "904ce6f8518b97a7628002ec15f7347f",
                "parameters": [
                    {
                        "name": "id_job",
                        "in": "path",
                        "description": "ID job offer yang ingin ditolak",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match berhasil ditolak dan dihapus"
                    },
                    "401": {
                        "description": "Belum login atau email belum diverifikasi"
                    },
                    "404": {
                        "description": "Match tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}